backend/pyproject.toml
[project]
name = "macrozero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.24",
    "tidb-vector>=0.0.15",
    "uvicorn>=0.35.0",
]
